cleanup(libscap): cleanup scap_stats_v2
