+++ userspace/libpman/src/capture.c
@@ -19,7 +19,7 @@ limitations under the License.
 #include <libpman.h> // before including scap so that handle_ppm_sc_mask() is not built
 #include <scap.h>
 #include "strlcpy.h"
-#include "../libscap/engine/modern_bpf/scap_modern_bpf.h"
+#include "../libscap/engine/modern_bpf/scap_modern_bpf_stats.h"
 
 /* This function should be idempotent, every time it is called it should enforce again the state */
 int pman_enforce_sc_set(bool *sc_set)
@@ -180,7 +180,7 @@ clean_print_stats:
 	return errno;
 }
 
-struct scap_stats_v2* pman_get_scap_stats_v2(void *scap_stats_v2_struct, uint32_t flags, uint32_t* nstats, int32_t* rc)
+void* pman_get_scap_stats_v2(void *scap_stats_v2_struct, uint32_t flags, uint32_t* nstats, int32_t* rc)
 {
 	struct scap_stats_v2 *stats = (struct scap_stats_v2 *)scap_stats_v2_struct;
 	*nstats = 0;
@@ -188,7 +188,6 @@ struct scap_stats_v2* pman_get_scap_stats_v2(void *scap_stats_v2_struct, uint32_
 	{
 		pman_print_error("pointer to scap_stats_v2 is empty");
 		*rc = errno;
-		return NULL;
 	}
 
 	if ((flags & PPM_SCAP_STATS_KERNEL_COUNTERS))
@@ -199,7 +198,6 @@ struct scap_stats_v2* pman_get_scap_stats_v2(void *scap_stats_v2_struct, uint32_
 		if(counter_maps_fd <= 0)
 		{
 			*rc = errno;
-			return stats;
 		}
 
 		/* We always take statistics from all the CPUs, even if some of them are not online.
@@ -208,8 +206,7 @@ struct scap_stats_v2* pman_get_scap_stats_v2(void *scap_stats_v2_struct, uint32_
 		for(uint32_t stat =  0;  stat < MODERN_BPF_MAX_KERNEL_COUNTERS_STATS; stat++)
 		{
 			stats[stat].type = STATS_VALUE_TYPE_U64;
-			stats[stat].flags = 0;
-			stats[stat].flags |= PPM_SCAP_STATS_KERNEL_COUNTERS;
+			stats[stat].flags = PPM_SCAP_STATS_KERNEL_COUNTERS;
 			stats[stat].value.u64 = 0;
 			strlcpy(stats[stat].name, modern_bpf_kernel_counters_stats_names[stat], STATS_NAME_MAX);
 		}
@@ -222,7 +219,6 @@ struct scap_stats_v2* pman_get_scap_stats_v2(void *scap_stats_v2_struct, uint32_
 				pman_print_error((const char *)error_message);
 				close(counter_maps_fd);
 				*rc = errno;
-				return stats;
 			}
 			stats[MODERN_BPF_N_EVTS].value.u64 += cnt_map.n_evts;
 			stats[MODERN_BPF_N_DROPS_BUFFER_TOTAL].value.u64 += cnt_map.n_drops_buffer;
@@ -234,8 +230,6 @@ struct scap_stats_v2* pman_get_scap_stats_v2(void *scap_stats_v2_struct, uint32_
 	// todo @incertum add libbpf stats for modern_bpf
 	*nstats = MODERN_BPF_MAX_KERNEL_COUNTERS_STATS;
 	*rc = SCAP_SUCCESS;
-	return stats;
-
 }
 
 int pman_get_n_tracepoint_hit(long *n_events_per_cpu)
