+++ userspace/libscap/engine/modern_bpf/scap_modern_bpf.c
@@ -247,7 +247,8 @@ int32_t scap_modern_bpf__get_stats(struct scap_engine_handle engine, OUT scap_st
 
 struct scap_stats_v2* scap_modern_bpf__get_stats_v2(struct scap_engine_handle engine, uint32_t flags, OUT uint32_t* nstats, OUT int32_t* rc)
 {
-	return pman_get_scap_stats_v2((void*)engine.m_handle->m_stats, flags, (void*)nstats, (void*)rc);
+	pman_get_scap_stats_v2((void*)engine.m_handle->m_stats, flags, (void*)nstats, (void*)rc);
+	return engine.m_handle->m_stats;
 }
 
 int32_t scap_modern_bpf__get_n_tracepoint_hit(struct scap_engine_handle engine, OUT long* ret)
