+++ userspace/libscap/engine/source_plugin/source_plugin_stats.h
@@ -1,5 +1,5 @@
 /*
-Copyright (C) 2022 The Falco Authors.
+Copyright (C) 2023 The Falco Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,21 +14,14 @@ See the License for the specific language governing permissions and
 limitations under the License.
 
 */
+
 #pragma once
 
-#include <stdint.h>
-#include "ringbuffer/devset.h"
-#include "scap_open.h"
-#include "scap_kmod_stats.h"
-#include "scap.h"
+typedef enum source_plugin_counters_stats {
+	N_EVTS = 0,
+	MAX_SOURCE_PLUGIN_COUNTERS_STATS,
+}source_plugin_counters_stats;
 
-struct kmod_engine
-{
-	struct scap_device_set m_dev_set;
-	char* m_lasterr;
-	interesting_ppm_sc_set curr_sc_set;
-	uint64_t m_api_version;
-	uint64_t m_schema_version;
-	bool capturing;
-	scap_stats_v2 m_stats[KMOD_MAX_KERNEL_COUNTERS_STATS];
+static const char * const source_plugin_counters_stats_names[] = {
+	[N_EVTS] = "n_evts",
 };
